Remove organise-specific forms
This commit removes the ActivityEditType and ActivityNewType forms, as these
were only used by the organise-related code. The Admin/ActivityEditType and
Admin/ActivityNewType forms have been moved a folder up, and all relevant code
has been changed accordingly.
