--- src/Form/Activity/Admin/ActivityNewType.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace App\Form\Activity\Admin;
-
-use Symfony\Component\Form\FormBuilderInterface;
-use Vich\UploaderBundle\Form\Type\VichImageType;
-
-class ActivityNewType extends ActivityEditType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        parent::buildForm($builder, $options);
-
-        $builder
-            ->add('imageFile', VichImageType::class, [
-                'label' => 'Image file',
-                'required' => true,
-                'allow_delete' => false,
-            ])
-        ;
-    }
-}
