+++ tests/Integration/Form/Activity/ActivityEditTypeTest.php
@@ -3,9 +3,17 @@
 namespace Tests\Integration\Form\Activity;
 
 use App\Entity\Activity\Activity;
+use App\Entity\Location\Location;
+use App\Entity\Security\LocalAccount;
 use App\Form\Activity\ActivityEditType;
-use DateTime;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Tools\SchemaTool;
+use Liip\TestFixturesBundle\Test\FixturesTrait;
+use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
+use Symfony\Component\Security\Guard\Token\PostAuthenticationGuardToken;
 
 /**
  * Class ActivityEditTypeTest.
@@ -14,6 +22,8 @@ use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
  */
 class ActivityEditTypeTest extends KernelTestCase
 {
+    use FixturesTrait;
+
     /**
      * {@inheritdoc}
      */
@@ -21,6 +31,24 @@ class ActivityEditTypeTest extends KernelTestCase
     {
         parent::setUp();
         self::bootKernel();
+
+        $firewallName = 'main';
+
+        $user = new LocalAccount();
+        $token = new PostAuthenticationGuardToken($user, $firewallName, ['ROLE_USER']);
+
+        /** @var TokenStorageInterface */
+        $storage = self::$container->get(TokenStorageInterface::class);
+        $storage->setToken($token);
+
+        // Get all database tables
+        $em = self::$container->get(EntityManagerInterface::class);
+        $cmf = $em->getMetadataFactory();
+        $classes = $cmf->getAllMetadata();
+
+        // Write all tables to database
+        $schema = new SchemaTool($em);
+        $schema->createSchema($classes);
     }
 
     /**
@@ -31,22 +59,38 @@ class ActivityEditTypeTest extends KernelTestCase
         parent::tearDown();
     }
 
-    public function testBuildForm(): void
+    public function testBindValidData(): void
     {
+        $location = new Location();
+        $location->setAddress('test');
+
         $type = new Activity();
-        $formData = [
-            'name' => 'test name',
+        $formdata = [
+            'name' => 'testname',
             'description' => 'test description',
-            'deadline' => new DateTime(),
-            'start' => new DateTime(),
-            'end' => new DateTime(),
+            'location' => $location,
+            'deadline' => 5,
+            'start' => 10,
+            'end' => 11,
+            'capacity' => 50,
+            'color' => 2,
         ];
 
         $formfactory = self::$container->get('form.factory');
         $form = $formfactory->create(ActivityEditType::class, $type);
 
-        $form->submit($formData);
+        $form->submit($formdata);
         self::assertTrue($form->isSynchronized());
         self::assertTrue($form->isSubmitted());
+    }
+
+    public function testConfigureOptions(): void
+    {
+        /** @var MockObject&OptionsResolver */
+        $resolver = $this->getMockBuilder("Symfony\Component\OptionsResolver\OptionsResolver")
+            ->disableOriginalConstructor()
+            ->getMock();
+        $resolver->expects(self::exactly(1))->method('setDefaults');
+        self::$container->get(ActivityEditType::class)->configureOptions($resolver);
     }
 }
